unit test for repo and service done

+++ src/main/java/com/example/IdeaHub/data/service/implementations/PublicationServiceImpl.java
@@ -1,7 +1,10 @@
-package com.example.IdeaHub.data.service;
+package com.example.IdeaHub.data.service.implementations;
 
-import com.example.IdeaHub.auth.model.ApplicationUser;
 import com.example.IdeaHub.auth.service.ApplicationUserDetails;
+import com.example.IdeaHub.data.model.Reviewers;
+import com.example.IdeaHub.data.repo.ReviewersRepo;
+import com.example.IdeaHub.data.service.interfaces.FileStorageService;
+import com.example.IdeaHub.data.service.interfaces.PublicationService;
 import com.example.IdeaHub.message.ResponseMessage;
 import com.example.IdeaHub.data.model.Publication;
 import com.example.IdeaHub.data.repo.PublicationRepo;
@@ -20,15 +23,18 @@ import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
-public class PublicationServiceImpl implements PublicationService{
+public class PublicationServiceImpl implements PublicationService {
 
-    private FileStorageService fileStorageService;
+    private final FileStorageService fileStorageService;
+
+    private final ReviewersRepo reviewersRepo;
 
     private final PublicationRepo publicationRepo;
 
     @Autowired
-    public PublicationServiceImpl(FileStorageService fileStorageService, PublicationRepo publicationRepo) {
+    public PublicationServiceImpl(FileStorageService fileStorageService, ReviewersRepo reviewersRepo, PublicationRepo publicationRepo) {
         this.fileStorageService = fileStorageService;
+        this.reviewersRepo = reviewersRepo;
         this.publicationRepo = publicationRepo;
     }
 
@@ -36,9 +42,10 @@ public class PublicationServiceImpl implements PublicationService{
 
         try{
             String message= "Upload Successful";
-            String authorId="";
+            String authorId;
             //getting current user
             Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+            System.out.println(principal);
             if(principal instanceof UserDetails){
                 authorId = ((ApplicationUserDetails) principal).getUserId();
                 publication.setAuthorId(authorId);
@@ -64,24 +71,28 @@ public class PublicationServiceImpl implements PublicationService{
     public ResponseEntity<ResponseMessage> updateReviewScore(String id) throws NoSuchElementException {
         int approvalThreshold = 3;
 
-
         try{
             Optional<Publication> publication = publicationRepo.findById(id);
 
-            int reviewScore = publication.get().getReviewScore() +1;
-            //increasing approval score by one
-            publication.get().setReviewScore( reviewScore);
-            publicationRepo.save(publication.get());
-
-            if(reviewScore >= approvalThreshold ){
-                int status  = updateApprovedStatus(publication.get());
-                if(status <= 0){
-                    return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ResponseMessage("Update Approval failed"));
-                }
+            if(publication.isEmpty()){
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
             }
+            else{
+                int reviewScore = publication.get().getReviewScore() +1;
+                //increasing approval score by one
+                publication.get().setReviewScore( reviewScore);
+                publicationRepo.save(publication.get());
 
-            String message = "Review Score updated successfully";
-            return ResponseEntity.status(HttpStatus.OK).body(new ResponseMessage(message));
+                if(reviewScore >= approvalThreshold ){
+                    int status  = updateApprovedStatus(publication.get());
+                    if(status <= 0){
+                        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ResponseMessage("Update Approval failed"));
+                    }
+                }
+
+                String message = "Review Score updated successfully";
+                return ResponseEntity.status(HttpStatus.OK).body(new ResponseMessage(message));
+            }
 
         } catch (NoSuchElementException e) {
             e.printStackTrace();
@@ -117,7 +128,7 @@ public class PublicationServiceImpl implements PublicationService{
 
     @Override
     public ResponseEntity<List<Publication>> getMyPublications(String authorId) {
-        List<Publication> publications = new ArrayList<>();
+        List<Publication> publications;
         try{
             publications = publicationRepo.findAllByAuthorId(authorId);
             return ResponseEntity.status(HttpStatus.OK).body(publications);
@@ -132,9 +143,9 @@ public class PublicationServiceImpl implements PublicationService{
     //these publications are approved by publication house
     @Override
     public ResponseEntity<List<Publication>> getApprovedPublications() {
-        List<Publication> publications = new ArrayList<>();
+        List<Publication> publications;
         try{
-            publicationRepo.findAllByApproved(true);
+            publications =publicationRepo.findAllByApproved(true);
             return ResponseEntity.status(HttpStatus.OK).body(publications);
         } catch (Exception e) {
             e.printStackTrace();
@@ -142,20 +153,43 @@ public class PublicationServiceImpl implements PublicationService{
         }
     }
 
+
+
+
+    //this queries publications by authorId and returns list of publication ids
+    //which are assigned to the author to review
+    public List<String> getPublicationIdsToReview(String authorId) {
+        List<Reviewers> reviewers = reviewersRepo.findAllByAuthorId(authorId);
+
+        List<String> publicationIds = new ArrayList<>();
+        //getting list of publication ids
+        for (Reviewers reviewers1: reviewers){
+            publicationIds.add(reviewers1.getPublicationId());
+        }
+        return publicationIds;
+    }
+
     @Override
     public ResponseEntity<List<Publication>> getPublicationsToReviewByAuthor() {
 
         //get it from filesToReview in applicationTable table
         List<Publication> publications = new ArrayList<>();
+        List<String> publicationIds ;
         try{
 
             //take author id from security context
-            String authorId="";
+            String authorId;
             //getting current user
             Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
             if(principal instanceof UserDetails){
                 authorId = ((ApplicationUserDetails) principal).getUserId();
-                publications = publicationRepo.findAllByReviewersEquals(authorId);
+
+                //getting publication ids
+                publicationIds = getPublicationIdsToReview(authorId);
+
+                for (String publicationId:publicationIds){
+                    publications.add(publicationRepo.findByUuid(publicationId));
+                }
                 return ResponseEntity.status(HttpStatus.OK).body(publications);
             }
 
@@ -168,29 +202,25 @@ public class PublicationServiceImpl implements PublicationService{
     }
 
     @Override
-    public ResponseEntity<ResponseMessage> assignReviewers(String id, List<String> reviewers) {
+    public ResponseEntity<ResponseMessage> assignReviewers(String publicationId, List<String> reviewers) {
         //checking if 5 reviewers are present
         if(reviewers.size() < 5){
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ResponseMessage("5 reviewers are mandatory"));
         }
         try{
             //first get the publication with that id
-            Optional<Publication> publication = publicationRepo.findById(id);
+            Optional<Publication> publication = publicationRepo.findById(publicationId);
             if (publication.isEmpty()){
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ResponseMessage("Publication with id "+ id+" was not found."));
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ResponseMessage("Publication with id "+ publicationId+" was not found."));
 
             }
             Publication publication1 = publication.get();
-            publication1.setReviewers(reviewers);
 
-
-//            //adding the publication id to the filesToReview section of author
-//            for(String reviewer:reviewers){
-//                //get the reviewers
-//            }
-
-            //now save the publication
-            publicationRepo.save(publication1);
+            //storing publication id and author ids to reviewers table
+            for (String reviewer: reviewers){
+                Reviewers reviewers1 = new Reviewers(publication1.getUuid(),reviewer);
+                reviewersRepo.save(reviewers1);
+            }
 
             return ResponseEntity.status(HttpStatus.OK).body(new ResponseMessage("Reviewers assigned successfully!"));
         } catch (Exception e) {
@@ -203,24 +233,30 @@ public class PublicationServiceImpl implements PublicationService{
     @SneakyThrows
     @Override
     public String storeFile(MultipartFile file)  {
-        return fileStorageService.store(file).getId();
+        return fileStorageService.storeFile(file).getId();
     }
 
     @Override
     public ResponseEntity<ResponseMessage> deletePublication(String id) {
         try{
-            Optional<Publication> publication = getPublication(id);
-            String fileId = publication.get().getFileId();
-            //deleting file associated with publication
-            int status = fileStorageService.deleteFile(fileId);
-            String message = "";
-            if(status <= 0){
-                message = "File associated with publication was not found";
-                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ResponseMessage(message));
+            Optional<Publication> publication = this.getPublication(id);
+
+            if(publication.isPresent()){
+                String fileId = publication.get().getFileId();
+                //deleting file associated with publication
+                int status = fileStorageService.deleteFile(fileId);
+                String message;
+                if(status <= 0){
+                    message = "File associated with publication was not found";
+                    return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ResponseMessage(message));
+                }
+                publicationRepo.deleteById(id);
+                message = "Publication deleted successfully";
+                return ResponseEntity.status(HttpStatus.OK).body(new ResponseMessage(message));
             }
-            publicationRepo.deleteById(id);
-            message = "Publication deleted successfully";
-            return ResponseEntity.status(HttpStatus.OK).body(new ResponseMessage(message));
+            else{
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            }
         } catch (Exception e) {
             e.printStackTrace();
             String message = e.getMessage();
@@ -243,4 +279,6 @@ public class PublicationServiceImpl implements PublicationService{
             return 0;
         }
     }
+
+
 }
